backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.127.0",
    "google-generativeai>=0.8.6",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
    "supabase>=2.0.0",
]
